% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{browse_url}
\alias{browse_url}
\title{Open a Web Browser with URL}
\usage{
browse_url(base, unencoded_query, encoded_query = "",
  browser = getOption("browser"), open_browser = interactive())
}
\arguments{
\item{base}{The URL prefix e.g. \code{https://google.com/search?q=}}

\item{unencoded_query}{An unencoded string that must be encoded with \code{\link[utils:URLencode]{utils::URLencode()}}.}

\item{encoded_query}{An encoded string that satisifies \code{\link[utils:URLencode]{utils::URLencode()}}.}

\item{browser}{a non-empty character string giving the name of the
    program to be used as the HTML browser.  It should be in the PATH,
    or a full path specified.  Alternatively, an \R function to be
    called to invoke the browser.

    Under Windows \code{NULL} is also allowed (and is the default), and
    implies that the file association mechanism will be used.
  }

\item{open_browser}{Should the URL be opened in a web browser.}
}
\description{
Open a web browser with a given URL.
}
\examples{

\dontrun{
# Query Google
browse_url("https://google.com/search?q=", "rstats is great")

# Generate URL for Google (do not open in a web browser)
browse_url("https://google.com/search?q=", "rstats is great",
           open_browser = FALSE)
}
}
\seealso{
\code{\link[utils:browseURL]{utils::browseURL()}}, \code{\link[utils:URLencode]{utils::URLencode()}}
}
