% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_func.R
\name{open_browser}
\alias{open_browser}
\title{Open a Web Browser with URL}
\usage{
open_browser(base, unencoded_query, encoded_query = "")
}
\arguments{
\item{base}{The URL prefix e.g. `https://google.com/search?q=`}

\item{unencoded_query}{An unencoded string that must be encoded with [utils::URLencode].}

\item{encoded_query}{An encoded string that satisifies [utils::URLencode].}
}
\description{
Open a web browser with a given URL.
}
\seealso{
[utils::browseURL], [utils::URLencode]
}
