% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_bing}
\alias{search_bing}
\title{Search Bing}
\usage{
search_bing(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_bing()} function searches \href{https://www.bing.com/}{Bing} using:
\verb{https://www.bing.com/search?q=<query>}
}
\examples{
# Search Bing
search_bing("Microsoft R")
}
\seealso{
Other search engines: 
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
