% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_bitbucket}
\alias{search_bitbucket}
\alias{search_bb}
\title{Search BitBucket}
\usage{
search_bitbucket(query = geterrmessage(), rlang = TRUE)

search_bb(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_bitbucket()} and \code{search_bb()} functions both search
\href{https://bitbucket.org}{BitBucket} using:
\verb{https://bitbucket.org/search?q=lang\\\%3Ar+<query>}
}
\details{
For additional details regarding \href{https://bitbucket.org}{BitBucket}'s
search interface please see:
\url{https://confluence.atlassian.com/bitbucket/code-search-in-bitbucket-873876782.html}
}
\examples{
# Search BitBucket for assertions
search_bitbucket("assertions")

# Search all languages on BitBucket for assertions
search_bitbucket("assertions", rlang = FALSE)
}
\seealso{
Other code repositories: 
\code{\link{search_github}()},
\code{\link{search_grep}()}
}
\concept{code repositories}
