% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_bluesky}
\alias{search_bluesky}
\title{Search BlueSky}
\usage{
search_bluesky(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_bluesky()} functions search
BlueSky using: \verb{https://bsky.app/search?q=\%23rstats+<query>}
}
\examples{
# Search BlueSky
search_bluesky("searcher")
}
\seealso{
Other community sites: 
\code{\link{search_mastodon}()},
\code{\link{search_posit_community}()},
\code{\link{search_stackoverflow}()},
\code{\link{search_twitter}()}
}
\concept{community sites}
