% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_duckduckgo}
\alias{search_duckduckgo}
\alias{search_ddg}
\title{Search DuckDuckGo}
\usage{
search_duckduckgo(query = geterrmessage(), rlang = TRUE)

search_ddg(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_duckduckgo()} and \code{search_ddg()} functions both search
\href{https://duckduckgo.com}{DuckDuckGo} using: \verb{https://duckduckgo.com/?q=<query>}
}
\examples{
# Search DuckDuckGo
search_duckduckgo("R language")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
