% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_ecosia}
\alias{search_ecosia}
\title{Search Ecosia}
\usage{
search_ecosia(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_ecosia()} function searches
Ecosia using: \verb{https://www.ecosia.org/search?q=<query>}
}
\details{
For additional details regarding Ecosia's search interface please see:
\url{https://support.ecosia.org/article/657-installing-ecosia-on-your-desktop-device}
}
\examples{
# Search Ecosia
search_ecosia("climate change R analysis")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
