% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_github}
\alias{search_github}
\alias{search_gh}
\title{Search GitHub}
\usage{
search_github(query = geterrmessage(), rlang = TRUE)

search_gh(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_github()} and \code{search_gh()} functions both search
\href{https://github.com}{GitHub} using:
\verb{https://github.com/search?q=<query>+language\\\%3Ar+type\\\%3Aissue&type=Issues}
}
\details{
For additional details regarding \href{https://github.com}{GitHub}'s
search interface please see:
\url{https://docs.github.com/en/enterprise-cloud@latest/search-github/getting-started-with-searching-on-github/about-searching-on-github}
and \url{https://docs.github.com/en/search-github/searching-on-github/searching-code/}
}
\examples{
# Search GitHub Issues for bivariate normal in the language:r
search_github("bivariate normal")

# Search all languages on GitHub Issues for bivariate normal
search_github("bivariate normal", rlang = FALSE)
}
\seealso{
Other code repositories: 
\code{\link{search_bitbucket}()},
\code{\link{search_grep}()}
}
\concept{code repositories}
