% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_google}
\alias{search_google}
\title{Search Google}
\usage{
search_google(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_google} function searches \href{https://www.google.com/}{Google} using:
\verb{https://www.google.com/search?q=<query>}
}
\examples{
# Search Google
search_google("r-project")

\dontrun{
# On error, automatically search the message on google
options(error = searcher("google"))
options(error = search_google)
}
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
