% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_grep}
\alias{search_grep}
\title{Search Grep.app}
\usage{
search_grep(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_grep()} function searches all public code on
\href{https://github.com}{GitHub} using \href{https://grep.app}{grep.app} by
querying: \verb{https://grep.app/search?q=<query-here>&filter[lang][0]=R}
}
\examples{
# Search R code on GitHub for numerical optimization
search_grep("optim")

# Search all code on GitHub for numerical optimization
search_grep("optim", rlang = FALSE)
}
\seealso{
Other code repositories: 
\code{\link{search_bitbucket}()},
\code{\link{search_github}()}
}
\concept{code repositories}
