% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_kagi}
\alias{search_kagi}
\title{Search Kagi}
\usage{
search_kagi(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_kagi()} function searches
Kagi using: \verb{https://kagi.com/search?q=<query>}
}
\details{
This is a paid search engine, and you will need to
sign up for an account to use it.
}
\examples{
# Search Kagi
search_kagi("advanced R programming")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
