% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_mastodon}
\alias{search_mastodon}
\title{Search Mastodon}
\usage{
search_mastodon(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_mastodon()} functions search
Mastodon using: \verb{https://mastodon.social/search?q=\%23rstats+<query>}
}
\examples{
# Search Mastodon
search_mastodon("searcher")
}
\seealso{
Other community sites: 
\code{\link{search_bluesky}()},
\code{\link{search_posit_community}()},
\code{\link{search_stackoverflow}()},
\code{\link{search_twitter}()}
}
\concept{community sites}
