% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_posit_community}
\alias{search_posit_community}
\alias{search_posit}
\title{Search Posit Community}
\usage{
search_posit_community(query = geterrmessage(), rlang = TRUE)

search_posit(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_posit_community()} and \code{search_posit()} functions both search
\href{https://forum.posit.co/}{Posit Community} using:
\verb{https://forum.posit.co/search?q=<query>}
}
\details{
For additional details regarding \href{https://forum.posit.co/}{Posit Community}'s
search interface please see the \href{https://discourse.org}{Discourse} API documentation:
\url{https://docs.discourse.org/#tag/Search}
}
\examples{
# Search Posit Community
search_posit_community("RStudio IDE")
}
\seealso{
Other community sites: 
\code{\link{search_bluesky}()},
\code{\link{search_mastodon}()},
\code{\link{search_stackoverflow}()},
\code{\link{search_twitter}()}
}
\concept{community sites}
