% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_qwant}
\alias{search_qwant}
\title{Search Qwant}
\usage{
search_qwant(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_qwant()} function searches
Qwant using: \verb{https://www.qwant.com/?q=<query>}
}
\examples{
# Search Qwant
search_qwant("Quarto")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_rseek}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
