% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_rseek}
\alias{search_rseek}
\title{Search Rseek}
\usage{
search_rseek(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_rseek()} function searches \href{https://rseek.org}{Rseek} using:
\verb{https://rseek.org/?q=<query>}
}
\examples{
# Search Rseek
search_rseek("searcher")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_startpage}()}
}
\concept{search engines}
