% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_stackoverflow}
\alias{search_stackoverflow}
\alias{search_so}
\title{Search StackOverflow}
\usage{
search_stackoverflow(query = geterrmessage(), rlang = TRUE)

search_so(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_stackoverflow()} and \code{search_so()} functions both search
\href{https://stackoverflow.com}{StackOverflow} using:
\verb{https://stackoverflow.com/search?q=\\\%5Br\\\%5D+<query>}
}
\details{
For additional details regarding \href{https://stackoverflow.com}{StackOverflow}'s
search interface please see:

\if{html}{\out{<div class="sourceCode">}}\preformatted{https://stackoverflow.com/help/searching
}\if{html}{\out{</div>}}
}
\examples{
# Search StackOverflow for Convolutions in the r tag
search_stackoverflow("convolutions")

# Search all languages on StackOverflow for convolutions
search_stackoverflow("convolutions", rlang = FALSE)
}
\seealso{
Other community sites: 
\code{\link{search_bluesky}()},
\code{\link{search_mastodon}()},
\code{\link{search_posit_community}()},
\code{\link{search_twitter}()}
}
\concept{community sites}
