% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_startpage}
\alias{search_startpage}
\alias{search_sp}
\title{Search Startpage}
\usage{
search_startpage(query = geterrmessage(), rlang = TRUE)

search_sp(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_startpage()} function searches
\href{https://startpage.com}{startpage} using:
\verb{https://startpage.com/do/dsearch?query=<query>}
}
\examples{
# Search startpage
search_startpage("VS Code")
}
\seealso{
Other search engines: 
\code{\link{search_bing}()},
\code{\link{search_brave}()},
\code{\link{search_duckduckgo}()},
\code{\link{search_ecosia}()},
\code{\link{search_google}()},
\code{\link{search_kagi}()},
\code{\link{search_qwant}()},
\code{\link{search_rseek}()}
}
\concept{search engines}
