% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{search_twitter}
\alias{search_twitter}
\title{Search Twitter}
\usage{
search_twitter(query = geterrmessage(), rlang = TRUE)
}
\arguments{
\item{query}{Contents of string to search. Default is the error message.}

\item{rlang}{Search for results written in R. Default is \code{TRUE}}
}
\value{
The generated search URL or an empty string.
}
\description{
The \code{search_twitter()} functions search
Twitter using: \verb{https://twitter.com/search?q=<query>}
}
\details{
For additional details regarding Twitter's
search interface please see:
\url{https://help.twitter.com/en/using-x/x-advanced-search}
}
\examples{
# Search Twitter
search_twitter("searcher")
}
\seealso{
Other community sites: 
\code{\link{search_posit_community}()},
\code{\link{search_stackoverflow}()}
}
\concept{community sites}
