% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct-functions.R
\name{searcher-defunct}
\alias{searcher-defunct}
\title{Defunct functions in \code{searcher}}
\description{
Functions listed below are no longer included in the \code{searcher} package
}
\details{
\itemize{
\item \code{search_ixquick()}: The function binding was removed as the search engine
name changed to "Startpage". Please use \code{\link[=search_startpage]{search_startpage()}}.
\item \code{search_rstudio_community()}: The function binding was removed as the
company name changed to "Posit". Please use \code{\link[=search_posit_community]{search_posit_community()}}.
\item \code{search_rscom()}: The function binding was removed as the
company name changed to "Posit". Please use \code{\link[=search_posit]{search_posit()}}.
}
}
