% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-functions.R
\name{searcher}
\alias{searcher}
\title{Generate a Searcher function for use with Error Handling}
\usage{
searcher(site, keyword = getOption("searcher.default_keyword"))
}
\arguments{
\item{site}{Name of site to search on. Supported options:
\code{"google"} (default), \code{"bing"}, \code{"duckduckgo"} or \code{"ddg"},
\code{"startpage"}  (formerly \code{"ixquick"}) or \code{"sp"},
\code{"qwant"}, \code{"rseek"}, \code{"brave"}, \code{"kagi"},
\code{"posit community"} (formerly \code{"rstudio community"}) or \code{"posit"},
\code{"twitter"} or \code{"x"}, \code{"bluesky"}, \code{"mastodon"}, \code{"stackoverflow"},
\code{"github"}, \code{"grep"}, \code{"bitbucket"},
\code{"chatgpt"}, \code{"claude"}, \code{"perplexity"},
\code{"mistral"} or \code{"le chat"}, \code{"bing copilot"} or \code{"copilot"}, and
\code{"grok"} or \code{"xai"}, \code{"meta ai"} or \code{"meta"}.}

\item{keyword}{Opt to search under different default terms.}
}
\description{
Constructs a function object that will search the last
R error message on search portals by opening a
browser.
}
\details{
This function acts as a closure. Thus, you will receive
a function back when only specifying the \code{site} parameter.
To call the function, add a second set of parentheses.
}
\section{Generic Error Search}{

The \code{searcher} function grabs the last error message and
tries to search it. This function will ensure that R language
is the primary search context.
}

\examples{
### Manually
searcher("google")()

\dontrun{
### Automatically
# On error, automatically search the message on google
options(error = searcher("google"))
}
}
