% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher-package.R
\docType{package}
\name{searcher-package}
\alias{searcher-package}
\alias{_PACKAGE}
\title{searcher: Query Search Engines Directly from R}
\description{
This package provides a search interface for R to search terms
on Google, Bing, DuckDuckGo, StackOverflow, GitHub, and BitBucket. Upon
searching, a browser window will open with the aforementioned search results.
}
\details{
Provides an interface to search engines via R.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
