% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher-package.R
\docType{package}
\name{searcher-package}
\alias{searcher-package}
\alias{_PACKAGE}
\title{searcher: Query Search Interfaces}
\description{
Provides a search interface to look up terms
    on 'Google', 'Bing', 'DuckDuckGo', 'Startpage', 'StackOverflow', 
    'RStudio Community', 'GitHub', and 'BitBucket'. Upon searching, a browser 
    window will open with the aforementioned search results.
}
\details{
Provides an interface to search engines via R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-assist/searcher}
  \item Report bugs at \url{https://github.com/r-assist/searcher/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458)

}
